frontend/requirements.txt
Flask
streamlit
pdfminer.six
python-docx
spacy
sentence-transformers
vaderSentiment
google-generativeai
pdfplumber
torch==2.6.0
